create separate file for relay related peer updates
